Week_2/Task_2/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-dotenv==1.1.0
langchain==0.3.26
wikipedia==1.4.0
langchain-openai==0.3.25
faiss-cpu==1.11.0
